Seperated home and edit components
