+++ client/src/App.css
@@ -62,6 +62,10 @@ canvas{
   width: 40%;
 }
 
+.homeimage{
+  max-width: 300px;
+}
+
 /* Gallery */
 .gallery {
   width: 90%;
